- Se ordenan las palabras por importancia y se escogen las primeras 200 y las 200 ultimas. - Modificada la generacion del BOW para que las palabras esten ordenadas por importancia.

+++ src/generatebowbaseline/ValueComparator.java
@@ -9,9 +9,9 @@ import java.util.Map;
  */
 public class ValueComparator implements Comparator<String> {
 
-    Map<String, Integer> base;
+    Map<String, Double> base;
 
-    public ValueComparator(Map<String, Integer> base) {
+    public ValueComparator(Map<String, Double> base) {
         this.base = base;
     }
